media/gpu/IP: Specify planes in ImageProcessor initialization
This is a follow up CL of crrev.com/c/1855142. The CL causes failures in
image_processor_test, because a layout of planes is not specified in
ImageProcessor initialization, and the adjusted planes are not exposed in
LibyuvImageProcessor. This CL fixes those mistakes.

Bug: 1005636
Test: image_processor_test on atlas
Change-Id: I481d0ca8d276721103030a9ba89d2b590050885d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1863065
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#706794}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 79dda366b71a5ace4dc6c5af89ede896b52e04b2