Speculative fix for null callback.
D3D11VideoDecoder::DoDecode checks |current_buffer_| to see if
there's a buffer being decoded.  If so, it assumes that there is
a callback for it.  Unfortunately, in some cases, this is wrong,
and it crashes.

NotifyError clears the callback without clearing |current_buffer_|,
and nothing else seems to.

This CL clears |current_buffer_| in NotifyError, to see if that's
the path by which the callback and the buffer are out of sync.

Bug: 1012464
Change-Id: I5d62ce07490de81bb1bfec7fa68a515efc416c11
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1922459
Commit-Queue: Frank Liberato <liberato@chromium.org>
Commit-Queue: Ted Meyer <tmathmeyer@chromium.org>
Auto-Submit: Frank Liberato <liberato@chromium.org>
Reviewed-by: Ted Meyer <tmathmeyer@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#716386}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7e52bbb29b8773229da4371ee31487c0f55a1ec7