Separate InitializeStaticGLBindings from InitializeGLOneOffImplementation
InitializeStaticGLBindings function is mainly for loading dlls which don't
call the GPU drivers or use the GPU hardware. This code separation is the
pre-work for the GPU watchdog to guard the functions that use the GPU
drivers or/and hardware only.

Bug:949839

Change-Id: I3f0a2f7eab6b17689dd088208e4298d2d410a52f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1854211
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Maggie Chen <magchen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#717220}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cb921fdfb771075cac846a04a6b21cd5735ea11a