media/gpu: return Optional<Fourcc> from Fourcc::FromVAFourCC()
Invalid Fourccs are dangerous by nature: they can be passed around just
like valid values and create bugs that manifest in a completely
different part of the code. To solve this, we plan to make it impossible
to create an invalid Fourcc instance.

This CL advances towards this goal by making Fourcc::FromVAFourCC()
return Optional<Fourcc>. If a value exists, it is guaranteed to be
valid, and if the conversion failed then nullopt is returned. This
forces the caller to check the returned value and will crash immediately
if the Fourcc is access in an attempt to pass the value around. All
callers of this method are also updated accordingly, with the required
checks added (these checks should have been there anyway).

BUG=1031082
BUG=b:143498574
TEST=Build completes on Eve.

Change-Id: I675ef7a4e1ec839d013f84f85152ed7ea3fd0406
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1957198
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: David Staessens <dstaessens@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#725060}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dbb1c0cdf908ad2fa435e9a22a470de0d7cc520e