media/gpu/chromeos: Duplicate fds of VideoFrame in CreateGpuMemoryBufferHandle()
VideoFrame might not have the same number of fds as the number of
planes. We duplicate the fd of the last plane of the VideoFrame
in CreateGpuMemoryBufferHandle() until the number of fds is the
same as the number of planes.

Bug: 1011281
Test: VDA test on hana
Change-Id: I6738e74ddc43f6c9457ba37beb44ee656866a633
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1978296
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#726698}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 233a69091a64031a8c9372ccac3e3e819a5bd903