Set the HOTWORD effect as a general effect
We used to treat the HOTWORD as a flag to enable DSP hardware hotwording for
Google Assistant. For scenario that we pin capture stream on internal
mic for board without usable DSP hotwording, we used not to enable it. With this
CL, we'll always enable the flag and rely on the passed in device_id to
determine if it is a DSP hotwording or not. It also sets the stream type to
SpeechRecognition to enable use case specific tuning on CRAS.

This change will depend on CRAS to also treat the mask flag as general HOTWORD.
Corresponding change is chromium:1994813 which needs to be fixed before this.

BUG=chromium:1020969
TEST=Build, deploy and check if CRAS creates the corresponding stream.

Change-Id: I4faeeeccfa11c409a7bfcc177d31c7223837b4b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997003
Commit-Queue: En-Shuo Hsu <enshuo@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#731855}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5c061177f464f44ade3ab77c03de44babee88260