media: Fix MediaFactory renderer factory selection logic
Currently when |use_media_player| is true, we'll add
FactoryType::kMediaPlayer as the base factory type. But we didn't set
|use_default_renderer_factory| to be false, so later we'll add
FactoryType::kDefault as the base factory type again, which overrides
the FactoryType::kMediaPlayer one we just added.

This CL makes sure we set |use_default_renderer_factory| to be false so
that this bug doesn't happen.

This function is still hard to read and error prone. In the future we
should improve this function further.

Bug: 1040456
Test: Manually tested.
Change-Id: Iade8a45426b3ff9835bb2b6cbc415f4df8a0cf23
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2005090
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#732665}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0d3ebf3b221b110686b4e1edf7fb76e65b59cb9f