media/gpu/V4L2: Fix bug causing V4L2 device poller to hang.
When we dequeue a buffer on the V4L2 device queue, we will immediately schedule
another poll on the device when there are more buffers left to dequeue. It is
possible however that we're notified of an event while there are no buffers
ready to dequeue. Trying to dequeue a buffer at this time would cause polling to
not be resumed, and we won't be notified when future buffers are ready to be
dequeued.

TEST=./video_encode_accelerator_unittest on hana with CL:1994545

BUG=b:147786740

Change-Id: Ieb2becd246ecbdb632ed930471cbc3fbfcf15e20
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2004354
Commit-Queue: David Staessens <dstaessens@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#732701}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ff0159c385b35fc0a153bb4d93f1f0c36d47216b