media/gpu/v4l2vda: fix EGL image size
The size of the EGL image is supposed to be the visible size, while the
buffer itself shall be imported with its own coded size and stride.

Before crrev.com/c/1982499, we did not pass any size information with
the buffer, so as a workaround the size was set to the coded size and
the buffer imported with its coded size into the GPU or DC, which
resulted in visible green lines on the right or lower end of the image
when coded_size != visible_size.

Now that we are importing the buffer using a NativePixmapHandle, we can
set the EGL image to be the correct visible size.

BUG=b:141579960
TEST=240p h.264 video plays in Youtube web without visible green line on
Hana.

Change-Id: Ia7631998b7e47c5f8d1001b26e5a154fdffebcb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2010588
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Auto-Submit: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#733539}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 35632d04c0496ea1db83325a4f54dde4b2a65337