media/gpu/v4l2(s)vda: harmonize buffer import code
Both decoders are supposed to do the same thing here, but small
differences in the code design could make the reader believe that they
don't. In particular, the IP is only supposed to be created in IMPORT
mode for both cases.

Make sure both decoders run the same version. We are going to reorganize
and factorize this code, but this little cleanup ensures it is safe to
do so.

BUG=None
TEST=Youtube desktop and Android both play on Hana and Kukui.

Change-Id: Ic572763b519f74337adc31714f6fa738d9bce194
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2011783
Reviewed-by: David Staessens <dstaessens@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Auto-Submit: Alexandre Courbot <acourbot@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#733875}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 59b895e8be9d8dff725df3ae885ac290f9b553e0